requirements.txt
Flask==2.2.3
Flask_Cors==4.0.0
numpy==1.24.2
pandas==2.0.3
psycopg2-binary==2.9.6
pygwalker==0.4.9.14
gunicorn==20.1.0
